[REFACTOR] presentation 패키지 정리
