+++ app/src/main/java/org/sopt/sample/presentation/MainActivity.kt
@@ -1,15 +1,16 @@
-package org.sopt.sample
+package org.sopt.sample.presentation
 
 import android.icu.lang.UCharacter.GraphemeClusterBreak.T
-import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.commit
 import androidx.fragment.app.replace
+import org.sopt.sample.R
 import org.sopt.sample.databinding.ActivityMainBinding
-import org.sopt.sample.fragment.HomeFragment
-import org.sopt.sample.fragment.SettingFragment
-import org.sopt.sample.fragment.UserFragment
+import org.sopt.sample.presentation.home.HomeFragment
+import org.sopt.sample.presentation.setting.SettingFragment
+import org.sopt.sample.presentation.user.UserFragment
 
 class MainActivity : AppCompatActivity() {
     private lateinit var binding: ActivityMainBinding
@@ -38,7 +39,7 @@ class MainActivity : AppCompatActivity() {
         }
     }
 
-    private inline fun <reified T: Fragment> setCurrentFragment() {
+    private inline fun <reified T : Fragment> setCurrentFragment() {
         supportFragmentManager.commit {
             replace<T>(R.id.container_home, T::class.java.canonicalName)
         }
