+++ app/src/main/java/org/sopt/sample/presentation/home/FollowerViewModel.kt
@@ -1,14 +1,8 @@
-package org.sopt.sample.data
+package org.sopt.sample.presentation.home
 
 import androidx.lifecycle.ViewModel
 import org.sopt.sample.data.remote.ResponseGetFollowerListDTO
 
 class FollowerViewModel : ViewModel() {
     val followerList = mutableListOf<ResponseGetFollowerListDTO.Follower>()
-
-    companion object {
-        const val PREF_FILE_NAME = "IN_SOPT"
-        const val PREF_USER_ID = "USER_ID"
-        const val PREF_USER_PWD = "USER_PWD"
-    }
 }
