+++ app/src/main/java/org/sopt/sample/presentation/setting/SettingFragment.kt
@@ -1,14 +1,13 @@
-package org.sopt.sample.fragment
+package org.sopt.sample.presentation.setting
 
 import android.content.Intent
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import org.sopt.sample.LoginActivity
-import org.sopt.sample.R
+import androidx.fragment.app.Fragment
 import org.sopt.sample.databinding.FragmentSettingBinding
+import org.sopt.sample.presentation.login.LoginActivity
 
 class SettingFragment : Fragment() {
     private var _binding: FragmentSettingBinding? = null
