+++ app/src/main/res/layout/activity_sign_up.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingHorizontal="40dp"
-    tools:context=".SignUpActivity">
+    tools:context=".presentation.signup.SignUpActivity">
 
     <TextView
         android:id="@+id/txt_title"
