+++ app/src/main/res/layout/fragment_home.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:paddingHorizontal="10dp"
     android:paddingBottom="10dp"
-    tools:context=".fragment.HomeFragment" >
+    tools:context=".presentation.home.HomeFragment" >
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/rv_follower"
